♻️ Separate typing and logic
[ADD] - Creation of cart-types.ts includes all types for the cart context.

[UPDATE] - Remove types from cart-context.tsx
